Adding client fiels

+++ src/main/java/ClientModel/Order.java
@@ -60,7 +60,7 @@ public class Order {
 		
 	}
 
-	public float getOrderId() {
+	public int getOrderId() {
 		return orderId;
 	}
